feat: 플레이어 생성하기 기능 구현
플레이어 생성하기 테스트
    - 정상적인 플레이어 생성
    - 잘못된 이름으로 플레잉 생성 시도

test 컨벤션 쓰려다가 커밋의 흐름상 feat로 판단함
