+++ README.md
@@ -422,9 +422,9 @@ pobi:--
 
 <h4>
 
-- [ ] 플레이어 생성하기
-    - [ ] 정상적인 플레이어 생성
-    - [ ] 잘못된 이름으로 플레이어 생성 시도
+- [x] 플레이어 생성하기
+    - [x] 정상적인 플레이어 생성
+    - [x] 잘못된 이름으로 플레이어 생성 시도
 
 
 - [ ] 플레이어 이동하기
