--- src/test/java/racingcar/player/RacingGamePlayerTest.java
@@ -1,8 +1,0 @@
-package racingcar.player;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-class RacingGamePlayerTest {
-
-}
-
